requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
path==16.6.0
Pillow==9.3.0
pyTelegramBotAPI==4.7.1
qrcode==7.3.1
requests==2.28.1
urllib3==1.26.13
